requirements.txt
uvloop
profilehooks
python-chess>=0.21.2
appnope>=0.1.0
bleach>=1.5.0
click>=6.7
cycler>=0.10.0
decorator>=4.1.2
entrypoints>=0.2.3
h5py>=2.7.1
html5lib>=0.9999999
ipykernel>=4.6.1
ipython>=6.2.1
ipython-genutils>=0.2.0
ipywidgets>=7.0.3
jedi>=0.11.0
Jinja2>=2.9.6
jsonschema>=2.6.0
jupyter>=1.0.0
jupyter-client>=5.1.0
jupyter-console>=5.2.0
jupyter-core>=4.3.0
Keras>=2.0.8
Markdown>=2.6.9
MarkupSafe>=1.0
matplotlib>=2.1.0
mistune>=0.7.4
nbconvert>=5.3.1
nbformat>=4.4.0
nose>=1.3.7
notebook>=5.2.0
numpy>=1.13.3
pandocfilters>=1.4.2
parso>=0.1.0
pexpect>=4.2.1
pickleshare>=0.7.4
prompt-toolkit>=1.0.15
protobuf>=3.4.0
ptyprocess>=0.5.2
Pygments>=2.2.0
pyparsing>=2.2.0
python-dateutil>=2.6.1
python-dotenv>=0.7.1
pytz>=2017.2
PyYAML>=3.12
pyzmq>=16.0.2
qtconsole>=4.3.1
scipy>=0.19.1
simplegeneric>=0.8.1
six>=1.11.0
tensorflow>=1.3.0
tensorflow-tensorboard>=0.1.8
terminado>=0.6
testpath>=0.3.1
tornado>=4.5.2
traitlets>=4.3.2
wcwidth>=0.1.7
Werkzeug>=0.12.2
widgetsnbextension>=3.0.6
